LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/canvas-dynamic-change-001-expected.txt


FAIL .flexbox 1 assert_equals:
<div class="flexbox column">
    <!-- The height=400 attribute makes the aspect ratio be 300x400. 300 from
         the fallback width of replaced elements. 400 from the attribute.
	 After stretching to 600px wide, the height should be 800px. -->
    <canvas id="canvas" data-expected-height="800" height="400"></canvas>
  </div>
height expected 800 but got 400
